feat: implement thread pool
- add functionality to pool threads
- implement a thread handler that takes up tasks from
the queue and executes them.
- refactor main.cpp
