+++ src/main.cpp
@@ -1,5 +1,6 @@
 #include "main.h"
 #include "Logger.h"
+#include "ThreadPoolSingleton.h"
 
 /**
  * Database Proxy main function
@@ -33,95 +34,6 @@ int main(int argc, char **argv) {
     return 0;
 }
 
-int updateProxyConfig(CProxySocket *socket, RESOLVED_PROXY_CONFIG configValue) {
-    TARGET_ENDPOINT_CONFIG targetEndpointConfig = {
-            .name = configValue.name,
-            .proxyPort = configValue.proxyPort,
-            .services = configValue.services};
-    socket->SetConfigValues(targetEndpointConfig);
-
-    return 0;
-};
-
-std::string updateProxyServers() {
-    std::vector<RESOLVED_PROXY_CONFIG> configValues = configSingleton.ResolveProxyServerConfigurations();
-    try {
-        for (auto value: configValues) {
-            updateProxyConfig(configSingleton.proxySocketsMap[value.name], value);
-        }
-    }
-    catch (std::exception &e) {
-        cout << e.what() << endl;
-        return "Error occurred when updating the configurations : " + std::string(e.what());
-    }
-    return "Configuration Updated Successfully!\n";
-}
-
-std::string updateConfiguration(std::string content) {
-    configSingleton.LoadConfigurationsFromString(std::move(content));
-    std::string response_content = updateProxyServers();
-    return response_content;
-}
-
-int initProtocolServers() {
-    std::vector<RESOLVED_PROTOCOL_CONFIG> protocolServerConfigValues = configSingleton.ResolveProtocolServerConfigurations();
-    for (const auto& value: protocolServerConfigValues) {
-        auto httpProtocolHandler = (CHttpProtocolHandler *) value.handler;
-        for (const auto& route: value.routes) {
-            httpProtocolHandler->RegisterHttpRequestHandler(
-                    route.path,
-                    HttpMethodEnumMap.find(route.method)->second,
-                    [route](const simple_http_server::HttpRequest &request) -> simple_http_server::HttpResponse {
-                        std::cout << "Starting request handler lambda :" << std::endl;
-
-                        ComputationContext context;
-                        /**
-                         * Set necessary shared variables and functions
-                         */
-                        context.Put("request", &request);
-                        context.Put("update_configuration", updateConfiguration);
-
-                        CommandDispatcher::Dispatch(route.request_handler, &context);
-                        auto response = std::any_cast<simple_http_server::HttpResponse *>(context.Get("response"));
-                        std::cout << (int) (response->status_code()) << std::endl;
-                        return *response;
-                    }
-            );
-        }
-        configSingleton.protocolSocketsMap[value.protocol_name] = new CProtocolSocket(value.protocol_port);
-        int protocolServer = startProtocolServer(
-                configSingleton.protocolSocketsMap[value.protocol_name],
-                value);
-        if (protocolServer < 0)
-            return protocolServer;
-    }
-
-    return 1;
-}
-
-int initProxyServers() {
-    std::vector<RESOLVED_PROXY_CONFIG> configValues = configSingleton.ResolveProxyServerConfigurations();
-    PipelineFactory pipelineFactory;
-    for (auto value: configValues) {
-        if (value.pipelineName == "ClickHouseLibuvPipeline") {
-            int proxy = startLibuvProxyServer(
-                    new LibuvProxySocket(value.proxyPort),
-                    pipelineFactory.GetLibuvProxyPipeline(value.pipelineName),
-                    value);
-            if (proxy < 0)
-                return proxy;
-        } else {
-            configSingleton.proxySocketsMap[value.name] = new CProxySocket(value.proxyPort);
-            int proxy = startProxyServer(
-                    configSingleton.proxySocketsMap[value.name],
-                    value);
-            if (proxy < 0)
-                return proxy;
-        }
-    }
-    return 0;
-}
-
 /**
  * Error handler
  */
@@ -145,118 +57,5 @@ void errorHandler(int sig) {
     exit(1);
 }
 
-int startLibuvProxyServer(
-        LibuvProxySocket *socket,
-        PipelineFunction<LibuvProxySocket> pipelineFunction,
-        RESOLVED_PROXY_CONFIG configValue
-) {
-    // Create proxies
-    std::string proxyName = configValue.name;
 
-    if (!(*socket).
-            SetPipeline(pipelineFunction)
-            ) {
-        cout << "Failed to set " << proxyName << " Pipeline ..!" <<
-             endl;
-        return -2;
-    }
-    CProxyHandler *proxyHandler = (CProxyHandler *) configValue.handler;
-    if (!(*socket).
-            SetHandler(proxyHandler)
-            ) {
-        cout << "Failed to set " << proxyName << " Handler ..!" <<
-             endl;
-        return -2;
-    }
 
-    TARGET_ENDPOINT_CONFIG targetEndpointConfig = {
-            .name = configValue.name,
-            .proxyPort = configValue.proxyPort,
-            .services = configValue.services};
-
-    if (!(*socket).
-            SetConfigValues(targetEndpointConfig)
-            ) {
-        cout << "Failed to set " << proxyName << " Config values ..!" <<
-             endl;
-        return -2;
-    }
-
-    if (!(*socket).
-            Start(proxyName)
-            ) {
-        cout << "Failed To Start " << proxyName << " Proxy Server ..!" <<
-             endl;
-        return -3;
-    }
-
-    return 0;
-}
-
-int startProxyServer(
-        CProxySocket *socket,
-        RESOLVED_PROXY_CONFIG configValue) {
-    // Create proxies
-    std::string proxyName = configValue.name;
-
-    // Setting up ClickHouse Proxy
-    PipelineFunction<CProxySocket> pipelineFunction = configValue.pipeline;
-    if (!(*socket).SetPipeline(pipelineFunction)) {
-        cout << "Failed to set " << proxyName << " Pipeline ..!" << endl;
-        return -2;
-    }
-    CProxyHandler *proxyHandler = (CProxyHandler *) configValue.handler;
-    if (!(*socket).SetHandler(proxyHandler)) {
-        cout << "Failed to set " << proxyName << " Handler ..!" << endl;
-        return -2;
-    }
-
-    TARGET_ENDPOINT_CONFIG targetEndpointConfig = {
-            .name = configValue.name,
-            .proxyPort = configValue.proxyPort,
-            .services = configValue.services};
-
-    if (!(*socket).SetConfigValues(targetEndpointConfig)) {
-        cout << "Failed to set " << proxyName << " Config values ..!" << endl;
-        return -2;
-    }
-
-    if (!(*socket).Start(proxyName)) {
-        cout << "Failed To Start " << proxyName << " Proxy Server ..!" << endl;
-        return -3;
-    }
-
-    return 0;
-}
-
-int startProtocolServer(
-        CProtocolSocket *socket,
-        RESOLVED_PROTOCOL_CONFIG configValue) {
-    // Create protocol
-    std::string protocolName = configValue.protocol_name;
-
-    // Setting up Protocol Pipeline
-    PipelineFunction<CProtocolSocket> pipelineFunction = configValue.pipeline;
-    if (!(*socket).SetPipeline(pipelineFunction)) {
-        cout << "Failed to set " << protocolName << " Pipeline ..!" << endl;
-        return -2;
-    }
-    CProtocolHandler *protocolHandler = (CProtocolHandler *) configValue.handler;
-    if (!(*socket).SetHandler(protocolHandler)) {
-        cout << "Failed to set " << protocolName << " Handler ..!" << endl;
-        return -2;
-    }
-
-    //    if (!(*socket).SetConfigValues(configValue))
-    //    {
-    //        cout << "Failed to set " << protocolName << " Config values ..!" << endl;
-    //        return -2;
-    //    }
-
-    if (!(*socket).Start()) {
-        cout << "Failed To Start " << protocolName << " Proxy Server ..!" << endl;
-        return -3;
-    }
-
-    return 0;
-}
